requirements.txt
fastapi==0.111.*
uvicorn[standard]==0.30.*
pydantic==2.*
yfinance==0.2.*
pandas==2.*
numpy>=1.26,<3
python-dotenv==1.0.*
requests>=2.31
google-genai>=0.4.0
openai==1.99.0
streamlit==1.36.*
plotly==5.*
httpx>=0.27
